Move functions to util files
